Ex6 still not working

+++ meta-compce460/recipes-kernel/irqgen-mod/files/irqgen_sysfs.c
@@ -107,7 +107,8 @@ IRQGEN_ATTR_RO(intr_handled);
 static ssize_t count_register_show(struct device *dev, struct device_attribute *attr, char *buf)
 {
     u32 val = irqgen_read_count();
-    return sprintf(buf, "%lu\n", val);
+    return sprintf(buf, "%u\n", val);  // Changed from %lu to %u
+
 }
 IRQGEN_ATTR_RO(count_register);
 
@@ -128,7 +129,7 @@ static ssize_t enabled_show(struct device *dev, struct device_attribute *attr, c
 {
     u32 regvalue = ioread32(IRQGEN_CTRL_REG);
     u8 val = FIELD_GET(IRQGEN_CTRL_REG_F_ENABLE, regvalue);
-    return sprintf(buf, "%u\n", val);
+    return sprintf(buf, "%lu\n", (unsigned long)val);
 }
 static ssize_t enabled_store(struct device *dev, struct device_attribute *attr, const char *buf, size_t count)
 {
